move bootstrap types into SDK
